+++ cx-portal/src/components/pages/TechnicalUserDetails/TechnicalUserDetailsContent.tsx
@@ -67,7 +67,7 @@ export default function TechnicalUserDetailsContent({
       copy: true,
     },
   ]
-  
+
   return (
     <section>
       <Button